Rename io to pyhat_io to avoid conflicts
