+++ libpyhat/pyhat_io/hcube.py
