+++ tests/derived/m3/conftest.py
@@ -1,7 +1,7 @@
 from unittest import mock
 
 import numpy as np
-from libpyhat.io import io_moon_mineralogy_mapper as iomm
+from libpyhat.pyhat_io import io_moon_mineralogy_mapper as iomm
 import pytest
 
 from . m3_wv import m3_wv
